move the `calculate_alpha` to the `bank_erosion.utils` module.
